Various changes.
I've added in a most of the commits on other branches; Ampers commits and ppipe
stuff (see below).

So many changes. Mainly self-evident; I've put the ppipe code in a submodule;
added a iterpipe submodule that contains code from iterpipes (not my own) and
a ShellPipe class; added a few more classes to the module, including PipeLines,
plus changes the way the regular Pipe works (ArgPipe); I've added a few more
commands to the command submodule. Finally, I've made an 'example' script that
shows a few of the things added (not threads though).

I've started calling my fork 'Piping' since it was confusing to have it share a
name with one of the main classes.